Changes OrtEnvironment so it can't be closed by users.
