+++ java/src/main/java/ai/onnxruntime/OnnxTensor.java
@@ -333,7 +333,7 @@ public class OnnxTensor implements OnnxValue {
    */
   static OnnxTensor createTensor(OrtEnvironment env, OrtAllocator allocator, Object data)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       TensorInfo info = TensorInfo.constructFromJavaArray(data);
       if (info.type == OnnxJavaType.STRING) {
         if (info.shape.length == 0) {
@@ -403,7 +403,7 @@ public class OnnxTensor implements OnnxValue {
    */
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, String[] data, long[] shape) throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       TensorInfo info =
           new TensorInfo(
               shape,
@@ -451,7 +451,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, FloatBuffer data, long[] shape)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       OnnxJavaType type = OnnxJavaType.FLOAT;
       return createTensor(type, allocator, data, shape);
     } else {
@@ -492,7 +492,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, DoubleBuffer data, long[] shape)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       OnnxJavaType type = OnnxJavaType.DOUBLE;
       return createTensor(type, allocator, data, shape);
     } else {
@@ -571,7 +571,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, ByteBuffer data, long[] shape, OnnxJavaType type)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       return createTensor(type, allocator, data, shape);
     } else {
       throw new IllegalStateException("Trying to create an OnnxTensor on a closed OrtAllocator.");
@@ -611,7 +611,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, ShortBuffer data, long[] shape)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       OnnxJavaType type = OnnxJavaType.INT16;
       return createTensor(type, allocator, data, shape);
     } else {
@@ -652,7 +652,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, IntBuffer data, long[] shape)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       OnnxJavaType type = OnnxJavaType.INT32;
       return createTensor(type, allocator, data, shape);
     } else {
@@ -693,7 +693,7 @@ public class OnnxTensor implements OnnxValue {
   static OnnxTensor createTensor(
       OrtEnvironment env, OrtAllocator allocator, LongBuffer data, long[] shape)
       throws OrtException {
-    if ((!env.isClosed()) && (!allocator.isClosed())) {
+    if (!allocator.isClosed()) {
       OnnxJavaType type = OnnxJavaType.INT64;
       return createTensor(type, allocator, data, shape);
     } else {
